+++ modules-misc/prov-dot/src/main/java/module-info.java
@@ -2,6 +2,6 @@ module prov.dot {
     requires org.apache.commons.io;
     requires org.apache.logging.log4j;
     requires prov.model;
-
+    requires jakarta.xml.bind;
     exports org.openprovenance.prov.dot;
 }