+++ modules-misc/prov-dot/src/test/java/module-info.java
@@ -6,6 +6,6 @@ module prov.dot.test {
     requires prov.dot;
     requires prov.model.test;
     requires prov.notation.test;
-
+    requires jakarta.xml.bind;
     exports org.openprovenance.prov.dot.test;
 }